+++ src/main/kotlin/ru/mipt/npm/gradle/internal/publishing.kt
@@ -110,7 +110,10 @@ internal fun Project.addGithubPublishing(
     githubOrg: String,
     githubProject: String
 ) {
-    if (requestPropertyOrNull("publishing.github") == "false") return
+    if (requestPropertyOrNull("publishing.github") == "false") {
+        logger.info("Skipping github publishing based on flag value")
+        return
+    }
 
     val githubUser: String = requestProperty("publishing.github.user")
     val githubToken: String = requestProperty("publishing.github.token")
@@ -137,7 +140,10 @@ internal fun Project.addGithubPublishing(
 }
 
 internal fun Project.addSpacePublishing(spaceRepo: String) {
-    if (requestPropertyOrNull("publishing.space") == "false") return
+    if (requestPropertyOrNull("publishing.space") == "false") {
+        logger.info("Skipping space publishing based on flag value")
+        return
+    }
 
     val spaceUser: String = requestProperty("publishing.space.user")
     val spaceToken: String = requestProperty("publishing.space.token")
@@ -165,7 +171,14 @@ internal fun Project.addSpacePublishing(spaceRepo: String) {
 }
 
 internal fun Project.addSonatypePublishing() {
-    if (requestPropertyOrNull("publishing.sonatype") == "false") return
+    if(isSnapshot()){
+        logger.info("Sonatype publishing skipped for dev version")
+        return
+    }
+    if (requestPropertyOrNull("publishing.sonatype") == "false") {
+        logger.info("Skipping sonatype publishing based on flag value")
+        return
+    }
 
     val sonatypeUser: String = requestProperty("publishing.sonatype.user")
     val sonatypePassword: String = requestProperty("publishing.sonatype.password")